requirements.txt
streamlit
pandas
numpy
scikit-learn
llama-index
llama_index.embeddings.langchain

langchain
openai
python-dotenv
requests
umap-learn
hdbscan
psycopg2-binary
nltk
plotly
sentence-transformers
matplotlib
-U langchain-community